admin can get job by companyID

+++ RecruitmentSystemAPI/Migrations/RecruitmentSystemContextModelSnapshot.cs
@@ -252,7 +252,7 @@ namespace RecruitmentSystemAPI.Migrations
 
                     b.Property<bool>("IsActive");
 
-                    b.Property<int>("JobRating");
+                    b.Property<float>("JobRating");
 
                     b.Property<string>("Province")
                         .HasMaxLength(50)
